+++ src/test/java/com/commandoby/stringCalculator/GetAnswerTests.java
@@ -1,5 +1,7 @@
 package com.commandoby.stringCalculator;
 
+import java.math.BigDecimal;
+
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -13,7 +15,7 @@ public class GetAnswerTests {
 	@ParameterizedTest
 	@CsvFileSource(resources = "/csv/GetAnswerTests.csv", numLinesToSkip = 1)
 	public void getAnswer_Test(String input, double expected) throws InvalidCharacterException, SubEquationException {
-		double actualAnswer = Application.getAnswer(input);
+		BigDecimal actualAnswer = Application.getAnswer(input);
 		Assertions.assertEquals(expected, actualAnswer);
 	}
